moved "eventInfoFragment" to schedule folder
