+++ app/src/main/java/org/hackillinois/android/view/home/HomeFragment.kt
@@ -24,10 +24,10 @@ import org.hackillinois.android.R
 import org.hackillinois.android.common.TimeInfo
 import org.hackillinois.android.database.entity.Event
 import org.hackillinois.android.view.MainActivity
-import org.hackillinois.android.view.eventinfo.EventInfoFragment
 import org.hackillinois.android.view.home.eventlist.EventClickListener
 import org.hackillinois.android.view.home.eventlist.EventsSection
 import org.hackillinois.android.view.home.eventlist.EventsSectionFragment
+import org.hackillinois.android.view.schedule.EventInfoFragment
 import org.hackillinois.android.viewmodel.HomeViewModel
 import java.lang.Exception
