+++ app/src/main/java/org/hackillinois/android/view/schedule/EventsAdapter.kt
@@ -12,7 +12,6 @@ import org.hackillinois.android.R
 import org.hackillinois.android.common.FavoritesManager
 import org.hackillinois.android.database.entity.Event
 import org.hackillinois.android.view.MainActivity
-import org.hackillinois.android.view.eventinfo.EventInfoFragment
 import org.hackillinois.android.view.home.eventlist.EventClickListener
 
 class EventsAdapter(
